Change API group
Changed API group from "ke.cattle.io" to "eks.cattle.io". The
original intention was for all future cloud provider controllers
to belong to the "ke.cattle.io" with the goal of changing
kontainer-engine from a package to an API group. Although good in
theory, the conventional pattern is one API group to one repo and
will be used to avoid complications. The CloudCredential field
has also been changed to AmazonCredentialSecret. CloudCredential
is a Rancher specific term and AmazonCredentialSecret better
describes what the field refers to.