Fix network fields detection
Prior, if the VirtualNetwork field on status was empty, it was
assumed that network fields were not provided. The controller
would then generate them. The VirtualNetwork field is always
empty at the beginning of an EKS cluster being created
regardless of if the network fields are provided or not. This
caused the controlplane and nodegroups to use subnets
belonging to different VPC's. This is invalid and causes an
error. Now, the provider is determined by evaluating whether
subnets have been provided or not. Also, nodegroups will now
use the networking fields on status which should match what
the cluster is using.