Reenqueue if already updating
Prior, it was assumed an object would be reenqueued when updating
its phase to "updating". If a cluster was already in "updating"
it would not be reenqueued because it was a noop. The reenqueue
is necessary for the onChange function to begin waiting on the
EKS cluster and nodegroups to finish updating. Now, if the
object is already in updating, it will be reenqueued.