Skip recording "cluster currently updating" error
Cluster updating error valid in that the controller should
retry but there is no actionable resolution as far as a
user is concerned. An update that has either been
initiated by eks-operator or another source is already in
progress. It is possible an update is not being
immediately reflected in upstream cluster state. The
config object will reenter the controller and then the
controller will wait for the update to finish.