Disallow creating cluster if one with same name exists
Before creating a new hosted EKS cluster, we first check the existing
clusters to ensure the name provided by the user is unique. If it is not
unique, then we wait for user to delete the cluster in Rancher or the
existing cluster in EKS.

Similarly, we check for other eksclusterconfigs with the same display
name in the case of duplicate creation before the EKS cluster is
created.

If the cluster never advanced past the "not created" status, then we
also should not delete any existing EKS clusters.