Add support for launch templates and spot instances
Everytime a new cluster is created, the operator creates a new launch
template that will be managed behind the scenes. The launch template is
tagged as managed by Rancher and it should not be modified or deleted.
This launch template is deleted when the cluster is deleted.

If a cluster is imported, then the operator looks for a launch template
with the predefined name. It is used as a managed launch template. When
an imported cluster is deleted, this template is not deleted because the
cluster is not removed from EKS in this case.

For each new node group, the user can use their own launch template. If
they do not provide a launch template, then the operator creates a new
version of the managed template specifically for this new node group. If
a node group is updated, then a new version of the managed launch
template is created and the old version is deleted.

Support for spot instances is also added. If a user specifies "true" for
spot instances, then the user should also specify a number of instance
types via spotInstanceTypes.