Fix possible conflict when editing during provision
If a user edits a cluster at the exact right (or wrong) time, then the
UpdateStatus call may produce a conflict and will error. When the
controller re-enters the create function, it will try to verify that a
cluster with the same name in EKS does not exist. It will find the one
that was created and error.

Instead, we retry the UpdateStatus call on conflict to protect against
this happening.