Only return keys that need to be deleted
Prior, keys that need to be updated would be returned by
GetKeysToDelete() and GetKeysToUpdate(). EKS differentiates
between labels/tagss that need to be deleted or updated. Keys that
need updates to their values should only be returned by
GetKeysToUpdate() and not GetKeysToDelete(). Now, there should
be no overlap between the keys returned by the two functions.