Always pass scaling config on update
Prior, if only a label update was necessary, the scaling config
fields were not included in the nodegroup config update. This
is considered invalid and will cause an error to be returned.
Now, scaling config fields are included whether they need to be
updated or not. An additional bug has been fixed where if some
labels needed to be deleted and other updated, only the updated
labels would be added to the config update.