Migrating to the AWS SDK for Go V2
The AWS SDK v1 (which is used by this project) is moving to maintenance
mode July 2024 and will be completely out of support July 2025.

With the long deprecation policy of Rancher we will still have the eks-operator
around when the AWS SDK v1 moves out of support.
We need to update the operator and migrate to v2 of the SDK.

Issue: https://github.com/rancher/eks-operator/issues/336